+++ EXAM/OOP1/src/ee/taltech/iti0202/store/shops/AllShop.java
@@ -2,7 +2,7 @@ package ee.taltech.iti0202.store.shops;
 
 
 public class AllShop extends Shop {
-    public AllShop(String name, Integer profit) {
+    public AllShop(String name, double profit) {
         super(name, profit);
     }
 }
